+++ back/src/main/java/com/back/domain/member/member/controller/AdmMemberController.java
@@ -1,6 +1,8 @@
 package com.back.domain.member.member.controller;
 
 
+import com.back.domain.member.member.service.MemberService;
+import com.back.global.rq.Rq;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -9,5 +11,9 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/auth")
 @RequiredArgsConstructor
 public class AdmMemberController {
+    private final MemberService memberService;
+    private final Rq rq;
+
+
 
 }
