+++ back/src/main/java/com/back/domain/member/member/controller/MemberAuthController.java
@@ -1,6 +1,9 @@
 package com.back.domain.member.member.controller;
 
-import com.back.domain.member.member.dto.*;
+import com.back.domain.member.member.dto.MenteeSignupRequest;
+import com.back.domain.member.member.dto.MentorSignupVerifyRequest;
+import com.back.domain.member.member.dto.MentorVerificationRequest;
+import com.back.domain.member.member.dto.LoginRequest;
 import com.back.domain.member.member.entity.Member;
 import com.back.domain.member.member.service.MemberService;
 import com.back.global.rq.Rq;
@@ -15,8 +18,8 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 @RequestMapping("/auth")
 @RequiredArgsConstructor
-@Tag(name = "MemberController", description = "회원 컨트롤러")
-public class  MemberController {
+@Tag(name = "MemberAuthController", description = "회원 인증 컨트롤러")
+public class MemberAuthController {
     private final MemberService memberService;
     private final Rq rq;
     private final EmailVerificationService emailVerificationService;
@@ -102,9 +105,9 @@ public class  MemberController {
         return new RsData<>("200-6", "토큰 갱신 성공");
     }
 
-    @DeleteMapping("/me")
+    @PostMapping("/me/withdraw")
     @Operation(summary = "회원 탈퇴")
-    public RsData<Void> deleteMember() {
+    public RsData<Void> withdrawMember() {
         Member currentUser = rq.getActor();
         memberService.deleteMember(currentUser);
 
@@ -115,35 +118,4 @@ public class  MemberController {
         return new RsData<>("200-7", "회원 탈퇴가 완료되었습니다.");
     }
 
-    @GetMapping("/me/mentee")
-    @Operation(summary = "멘티 마이페이지 조회")
-    public RsData<MenteeMyPageResponse> getMenteeMyPage() {
-        Member currentUser = rq.getActor();
-        MenteeMyPageResponse response = memberService.getMenteeMyPage(currentUser);
-        return new RsData<>("200-9", "멘티 정보 조회 성공", response);
-    }
-
-    @PutMapping("/me/mentee")
-    @Operation(summary = "멘티 정보 수정")
-    public RsData<Void> updateMentee(@RequestBody MenteeUpdateRequest request) {
-        Member currentUser = rq.getActor();
-        memberService.updateMentee(currentUser, request);
-        return new RsData<>("200-10", "멘티 정보 수정 성공");
-    }
-
-    @GetMapping("/me/mentor")
-    @Operation(summary = "멘토 마이페이지 조회")
-    public RsData<MentorMyPageResponse> getMentorMyPage() {
-        Member currentUser = rq.getActor();
-        MentorMyPageResponse response = memberService.getMentorMyPage(currentUser);
-        return new RsData<>("200-11", "멘토 정보 조회 성공", response);
-    }
-
-    @PutMapping("/me/mentor")
-    @Operation(summary = "멘토 정보 수정")
-    public RsData<Void> updateMentor(@RequestBody MentorUpdateRequest request) {
-        Member currentUser = rq.getActor();
-        memberService.updateMentor(currentUser, request);
-        return new RsData<>("200-12", "멘토 정보 수정 성공");
-    }
 }
